added tokio dep + restructured parsing for adding async
